+++ changelog.md
@@ -7,6 +7,7 @@
 - Adds auto targeting to the trident
 - Adds a toggle to disable auto targeting (shift + mode switch key)
 - Adds a sweeping edge upgrade to the sword and aiot
+- Adds a swift sneak upgrade to the leggings
 
 ### Fixes
 - Fixes the fortune tooltip being incorrect on the quarry